# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library API",
    "description": "A REST API for managing a library's books and authors, built with Node.js, Express, and MongoDB (Mongoose).\n\n**How to test:** click an endpoint, press *Try it out*, then *Execute*. Requests go to the host serving this page.\n1. `POST /api/authors` - create an author and copy the `_id` from the response.\n2. `POST /api/books` - paste that `_id` into the `author` field and create a book (ISBN must be unique).\n3. `GET /api/books` - confirm the author is populated.\n4. `PUT /api/books/{id}` - update it (partial bodies are fine).\n5. `DELETE /api/books/{id}` - remove it.\n\n**Authentication:** GET routes are public. POST, PUT and DELETE require GitHub OAuth login. Open `/login` in the same browser first (Swagger UI cannot run the OAuth redirect itself); the session cookie is then sent automatically with *Try it out* requests. Use `/logout` to end the session.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server (the same host that is serving this page)"
    }
  ],
  "tags": [
    {
      "name": "Authors",
      "description": "Operations on the /api/authors collection"
    },
    {
      "name": "Books",
      "description": "Operations on the /api/books collection"
    },
    {
      "name": "System",
      "description": "Service status"
    },
    {
      "name": "Authentication",
      "description": "GitHub OAuth login/logout"
    }
  ],
  "paths": {
    "/api/authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get all authors",
        "responses": {
          "200": {
            "description": "List of authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Author"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Create a new author (requires login)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              },
              "example": {
                "name": "George Orwell",
                "bio": "English novelist and essayist.",
                "birthYear": 1903,
                "nationality": "British",
                "website": "https://www.orwellfoundation.com"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Author"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/authors/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "MongoDB ObjectId - the `_id` returned when the record was created."
        }
      ],
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get a single author",
        "responses": {
          "200": {
            "description": "Author found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Author"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Authors"
        ],
        "summary": "Update an existing author (requires login)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorUpdate"
              },
              "example": {
                "birthYear": 1904
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Author"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "summary": "Delete an author (requires login)",
        "responses": {
          "200": {
            "description": "Author deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Author"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books (author info populated)",
        "responses": {
          "200": {
            "description": "List of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 1
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Book"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Create a new book (requires login)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              },
              "example": {
                "title": "1984",
                "author": "PASTE_AUTHOR_ID_FROM_POST_/api/authors",
                "isbn": "978-0-452-28423-4",
                "genre": "Dystopian Fiction",
                "publishedYear": 1949,
                "pages": 328,
                "rating": 4.8,
                "description": "A dystopian social science fiction novel.",
                "inStock": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error, or referenced author does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "The `author` must be the `_id` of an existing author. Create one first with `POST /api/authors`.",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/books/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "MongoDB ObjectId - the `_id` returned when the record was created."
        }
      ],
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a single book",
        "responses": {
          "200": {
            "description": "Book found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update an existing book (partial updates allowed) (requires login)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              },
              "example": {
                "rating": 4.9,
                "inStock": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error, invalid ID, or referenced author does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book (requires login)",
        "responses": {
          "200": {
            "description": "Book deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health check (includes database connection state)",
        "responses": {
          "200": {
            "description": "API is up and the database is connected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "database": {
                      "type": "string",
                      "example": "connected"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "API is up but the database is not connected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "status": {
                      "type": "string",
                      "example": "degraded"
                    },
                    "database": {
                      "type": "string",
                      "example": "disconnected"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in with GitHub (OAuth)",
        "description": "Open this URL in the browser (not via 'Try it out'). Redirects to GitHub, then back to /api-docs. Creates a user record on first login.",
        "responses": {
          "302": {
            "description": "Redirect to GitHub authorization page"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "GitHub OAuth callback",
        "description": "Called by GitHub after authorization. Not meant to be called manually.",
        "responses": {
          "302": {
            "description": "Redirects to /api-docs on success"
          },
          "401": {
            "description": "GitHub login failed"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Check current login status",
        "description": "Returns whether the current session is logged in.",
        "responses": {
          "200": {
            "description": "Login status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "loggedIn": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "githubId": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out",
        "description": "Destroys the session.",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Logged out"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66f0c2a1e4b0a1a2b3c4d5e6"
          },
          "name": {
            "type": "string",
            "example": "George Orwell"
          },
          "bio": {
            "type": "string",
            "example": "English novelist and essayist."
          },
          "birthYear": {
            "type": "integer",
            "example": 1903
          },
          "nationality": {
            "type": "string",
            "example": "British"
          },
          "website": {
            "type": "string",
            "example": "https://www.orwellfoundation.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthorInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Author's full name (required)"
          },
          "bio": {
            "type": "string",
            "description": "Short biography"
          },
          "birthYear": {
            "type": "integer",
            "description": "1000 - current year"
          },
          "nationality": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "description": "Must be a valid URL if provided"
          }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66f0c3b2e4b0a1a2b3c4d5e7"
          },
          "title": {
            "type": "string",
            "example": "1984"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "isbn": {
            "type": "string",
            "example": "978-0-452-28423-4"
          },
          "genre": {
            "type": "string",
            "example": "Dystopian Fiction"
          },
          "publishedYear": {
            "type": "integer",
            "example": 1949
          },
          "pages": {
            "type": "integer",
            "example": 328
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "description": {
            "type": "string"
          },
          "inStock": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookInput": {
        "type": "object",
        "required": [
          "title",
          "author",
          "isbn",
          "genre",
          "publishedYear",
          "pages"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string",
            "description": "Must reference an existing Author's _id",
            "example": "PASTE_AUTHOR_ID_FROM_POST_/api/authors"
          },
          "isbn": {
            "type": "string",
            "description": "Must be unique"
          },
          "genre": {
            "type": "string"
          },
          "publishedYear": {
            "type": "integer",
            "description": "1000 - current year"
          },
          "pages": {
            "type": "integer",
            "description": "Must be >= 1"
          },
          "rating": {
            "type": "number",
            "description": "0 - 5, defaults to 0"
          },
          "description": {
            "type": "string"
          },
          "inStock": {
            "type": "boolean",
            "description": "Defaults to true"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Descriptive error message"
          }
        }
      },
      "AuthorUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Author's full name (required)"
          },
          "bio": {
            "type": "string",
            "description": "Short biography"
          },
          "birthYear": {
            "type": "integer",
            "description": "1000 - current year"
          },
          "nationality": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "description": "Must be a valid URL if provided"
          }
        },
        "description": "Any subset of fields may be sent."
      },
      "BookUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string",
            "description": "Must reference an existing Author's _id"
          },
          "isbn": {
            "type": "string",
            "description": "Must be unique"
          },
          "genre": {
            "type": "string"
          },
          "publishedYear": {
            "type": "integer",
            "description": "1000 - current year"
          },
          "pages": {
            "type": "integer",
            "description": "Must be >= 1"
          },
          "rating": {
            "type": "number",
            "description": "0 - 5, defaults to 0"
          },
          "description": {
            "type": "string"
          },
          "inStock": {
            "type": "boolean",
            "description": "Defaults to true"
          }
        },
        "description": "Any subset of fields may be sent."
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after logging in at /login (GitHub OAuth). The browser sends it automatically."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - you must log in first (visit /login)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Unauthorized: you must log in first. Visit /login"
            }
          }
        }
      }
    }
  }
}